added a postInit step to the SharedData base class, allowing Client and Server nested classes to have initialization steps after the shared data ptr is set. also, some other silly changes

+++ include/Framework/Graphics/Mesh.hpp
@@ -359,8 +359,8 @@ public:
         glEnableVertexAttribArray(0);
 
         // keep it thread-accurate
-        m_GPUVertCount = (m_verts.size() * sizeof(T)) / m_singleVertexSize;
-        m_GPUIndicesCount = m_indices.size();
+        m_GPUVertCount = uint32_t(m_verts.size() * sizeof(T)) / m_singleVertexSize;
+        m_GPUIndicesCount = (uint32_t)m_indices.size();
     };
 
     void pushInstancesToGPU() {
